requirements.txt
aioredis==1.0.0
asgiref==2.2.0
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
cffi==1.11.5
channels==2.0.2
channels-redis==2.1.1
constantly==15.1.0
daphne==2.1.0
Django==2.0.3
hiredis==0.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
msgpack==0.5.6
py-bcrypt==0.4
pycparser==2.18
python-dateutil==2.7.2
pytz==2018.3
six==1.11.0
Twisted==17.9.0
txaio==2.9.0
zope.interface==4.4.3
